src/backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Pillow==10.2.0  # for handling image uploads
django-filter==24.1  # for advanced filtering
